flag to turn off automatic save on end of session
The default value make it so there is no behavior change.
saveState is emitted every time there is a response. Normally it won't
save unless saveOnResponse is true, but any kind of tell request by
definition also has shouldEndSession true, and will therefore save to
database. We might not want that, so this flag allows you to control
whether or not save is called in the case of shouldEndSession.