Adds ability to set own directives
Developing with the Echo Show requires the use of directives in the response object. In the current state of the SDK, a developer must call this.context.succeed with a manually created JSON object. The structure of this object is the same as the one already created in the SDK - the only addition is the Display.RenderTemplate and optional Hint directives. This change allows a developer to just set these directives by setting this.handler.directives, taking advantage of all of the other features of the SDK.