Improve isObject check
Source: https://toddmotto.com/understanding-javascript-types-and-reliable-type-checking/#true-object-types

This is a slightly better check for objects as it allows ES2015 Proxy
objects as well. Proxies behave like objects, but the `constructor`
property of a Proxy is Function, instead of Object.

Proxies can be useful in implementing, for example, a catch-all handler
function for an Alexa Skill.